Added Circuit Breaker pattern with Resilience4j
Circuit Breaker is a pattern used to detect failures and prevent them from recurring constantly, preventing one microservice from continuing to make requests to another when there is a high probability that the process will fail.
Circuit Breaker shuts down the communication of a specific service with the others when it fails.
The circuit breaker must be in the microservice that communicates with the others, in this case, the user microservice.